refactor: diet-sessions featureにrepositoryレイヤーを追加
