+++ web/src/features/diet-sessions/server/loaders/get-previous-diet-session.ts
@@ -1,7 +1,7 @@
-import { createAdminClient } from "@mirai-gikai/supabase";
 import { unstable_cache } from "next/cache";
 import { CACHE_TAGS } from "@/lib/cache-tags";
 import type { DietSession } from "../../shared/types";
+import { findPreviousDietSession } from "../repositories/diet-session-repository";
 import { getActiveDietSession } from "./get-active-diet-session";
 
 /**
@@ -22,23 +22,7 @@ export async function getPreviousDietSession(): Promise<DietSession | null> {
 
 const _getCachedPreviousDietSession = unstable_cache(
   async (activeStartDate: string): Promise<DietSession | null> => {
-    const supabase = createAdminClient();
-
-    // アクティブなセッションより古いセッションを取得（start_dateで比較）
-    const { data: previousSession, error: previousError } = await supabase
-      .from("diet_sessions")
-      .select("*")
-      .lt("start_date", activeStartDate)
-      .order("start_date", { ascending: false })
-      .limit(1)
-      .maybeSingle();
-
-    if (previousError) {
-      console.error("Failed to fetch previous diet session:", previousError);
-      return null;
-    }
-
-    return previousSession;
+    return findPreviousDietSession(activeStartDate);
   },
   ["previous-diet-session"],
   {
